+++ apps/dash-oil-and-gas/requirements.txt
@@ -1,3 +1,3 @@
-pandas==0.24.2
-dash==1.12.0
-gunicorn==19.9.0
+dash==2.4.1
+pandas==1.4.2
+gunicorn==20.1.0
