+++ apps/dash-oil-and-gas/runtime.txt
@@ -1,0 +1,1 @@
+python-3.8.0
