+++ apps/dash-oil-and-gas/utils/data.py
@@ -1,0 +1,25 @@
+import urllib.request
+import pickle
+import pandas as pd
+import datetime as dt
+from constants import DATA_PATH
+
+
+# Download pickle file
+urllib.request.urlretrieve(
+    "https://raw.githubusercontent.com/plotly/datasets/master/dash-sample-apps/dash-oil-and-gas/data/points.pkl",
+    DATA_PATH.joinpath("points.pkl"),
+)
+points = pickle.load(open(DATA_PATH.joinpath("points.pkl"), "rb"))
+
+# Load data
+df = pd.read_csv(
+    "https://github.com/plotly/datasets/raw/master/dash-sample-apps/dash-oil-and-gas/data/wellspublic.csv",
+    low_memory=False,
+)
+df["Date_Well_Completed"] = pd.to_datetime(df["Date_Well_Completed"])
+df = df[df["Date_Well_Completed"] > dt.datetime(1960, 1, 1)]
+
+trim = df[["API_WellNo", "Well_Type", "Well_Name"]]
+trim.index = trim["API_WellNo"]
+dataset = trim.to_dict(orient="index")
