+++ apps/dash-oil-and-gas/utils/figures.py
@@ -1,0 +1,251 @@
+from utils.data import df, dataset
+from utils.helper_functions import (
+    filter_dataframe,
+    produce_individual,
+    produce_aggregate,
+)
+from constants import layout, WELL_TYPES, WELL_COLORS
+import copy
+
+import datetime as dt
+
+
+def make_main_figure(
+    well_statuses, well_types, year_slider, selector, main_graph_layout
+):
+
+    dff = filter_dataframe(df, well_statuses, well_types, year_slider)
+
+    traces = []
+    for well_type, dfff in dff.groupby("Well_Type"):
+        trace = dict(
+            type="scattermapbox",
+            lon=dfff["Surface_Longitude"],
+            lat=dfff["Surface_latitude"],
+            text=dfff["Well_Name"],
+            customdata=dfff["API_WellNo"],
+            name=WELL_TYPES[well_type],
+            marker=dict(size=4, opacity=0.6),
+        )
+        traces.append(trace)
+
+    # relayoutData is None by default, and {'autosize': True} without relayout action
+    if main_graph_layout is not None and selector is not None and "locked" in selector:
+        if "mapbox.center" in main_graph_layout.keys():
+            lon = float(main_graph_layout["mapbox.center"]["lon"])
+            lat = float(main_graph_layout["mapbox.center"]["lat"])
+            zoom = float(main_graph_layout["mapbox.zoom"])
+            layout["mapbox"]["center"]["lon"] = lon
+            layout["mapbox"]["center"]["lat"] = lat
+            layout["mapbox"]["zoom"] = zoom
+
+    figure = dict(data=traces, layout=layout)
+    return figure
+
+
+def make_individual_figure(main_graph_hover):
+
+    layout_individual = copy.deepcopy(layout)
+
+    if main_graph_hover is None:
+        main_graph_hover = {
+            "points": [
+                {"curveNumber": 4, "pointNumber": 569, "customdata": 31101173130000}
+            ]
+        }
+
+    chosen = [point["customdata"] for point in main_graph_hover["points"]]
+    index, gas, oil, water = produce_individual(chosen[0])
+
+    if index is None:
+        annotation = dict(
+            text="No data available",
+            x=0.5,
+            y=0.5,
+            align="center",
+            showarrow=False,
+            xref="paper",
+            yref="paper",
+        )
+        layout_individual["annotations"] = [annotation]
+        data = []
+    else:
+        data = [
+            dict(
+                type="scatter",
+                mode="lines+markers",
+                name="Gas Produced (mcf)",
+                x=index,
+                y=gas,
+                line=dict(shape="spline", smoothing=2, width=1, color="#fac1b7"),
+                marker=dict(symbol="diamond-open"),
+            ),
+            dict(
+                type="scatter",
+                mode="lines+markers",
+                name="Oil Produced (bbl)",
+                x=index,
+                y=oil,
+                line=dict(shape="spline", smoothing=2, width=1, color="#a9bb95"),
+                marker=dict(symbol="diamond-open"),
+            ),
+            dict(
+                type="scatter",
+                mode="lines+markers",
+                name="Water Produced (bbl)",
+                x=index,
+                y=water,
+                line=dict(shape="spline", smoothing=2, width=1, color="#92d8d8"),
+                marker=dict(symbol="diamond-open"),
+            ),
+        ]
+        layout_individual["title"] = dataset[chosen[0]]["Well_Name"]
+
+    figure = dict(data=data, layout=layout_individual)
+    return figure
+
+
+def make_aggregate_figure(well_statuses, well_types, year_slider, main_graph_hover):
+
+    layout_aggregate = copy.deepcopy(layout)
+
+    if main_graph_hover is None:
+        main_graph_hover = {
+            "points": [
+                {"curveNumber": 4, "pointNumber": 569, "customdata": 31101173130000}
+            ]
+        }
+
+    chosen = [point["customdata"] for point in main_graph_hover["points"]]
+    well_type = dataset[chosen[0]]["Well_Type"]
+    dff = filter_dataframe(df, well_statuses, well_types, year_slider)
+
+    selected = dff[dff["Well_Type"] == well_type]["API_WellNo"].values
+    index, gas, oil, water = produce_aggregate(selected, year_slider)
+
+    data = [
+        dict(
+            type="scatter",
+            mode="lines",
+            name="Gas Produced (mcf)",
+            x=index,
+            y=gas,
+            line=dict(shape="spline", smoothing="2", color="#F9ADA0"),
+        ),
+        dict(
+            type="scatter",
+            mode="lines",
+            name="Oil Produced (bbl)",
+            x=index,
+            y=oil,
+            line=dict(shape="spline", smoothing="2", color="#849E68"),
+        ),
+        dict(
+            type="scatter",
+            mode="lines",
+            name="Water Produced (bbl)",
+            x=index,
+            y=water,
+            line=dict(shape="spline", smoothing="2", color="#59C3C3"),
+        ),
+    ]
+    layout_aggregate["title"] = "Aggregate: " + WELL_TYPES[well_type]
+
+    figure = dict(data=data, layout=layout_aggregate)
+    return figure
+
+
+def make_pie_figure(well_statuses, well_types, year_slider):
+
+    layout_pie = copy.deepcopy(layout)
+
+    dff = filter_dataframe(df, well_statuses, well_types, year_slider)
+
+    selected = dff["API_WellNo"].values
+    index, gas, oil, water = produce_aggregate(selected, year_slider)
+
+    aggregate = dff.groupby(["Well_Type"]).count()
+
+    data = [
+        dict(
+            type="pie",
+            labels=["Gas", "Oil", "Water"],
+            values=[sum(gas), sum(oil), sum(water)],
+            name="Production Breakdown",
+            text=[
+                "Total Gas Produced (mcf)",
+                "Total Oil Produced (bbl)",
+                "Total Water Produced (bbl)",
+            ],
+            hoverinfo="text+value+percent",
+            textinfo="label+percent+name",
+            hole=0.5,
+            marker=dict(colors=["#fac1b7", "#a9bb95", "#92d8d8"]),
+            domain={"x": [0, 0.45], "y": [0.2, 0.8]},
+        ),
+        dict(
+            type="pie",
+            labels=[WELL_TYPES[i] for i in aggregate.index],
+            values=aggregate["API_WellNo"],
+            name="Well Type Breakdown",
+            hoverinfo="label+text+value+percent",
+            textinfo="label+percent+name",
+            hole=0.5,
+            marker=dict(colors=[WELL_COLORS[i] for i in aggregate.index]),
+            domain={"x": [0.55, 1], "y": [0.2, 0.8]},
+        ),
+    ]
+    layout_pie["title"] = "Production Summary: {} to {}".format(
+        year_slider[0], year_slider[1]
+    )
+    layout_pie["font"] = dict(color="#777777")
+    layout_pie["legend"] = dict(
+        font=dict(color="#CCCCCC", size="10"), orientation="h", bgcolor="rgba(0,0,0,0)"
+    )
+
+    figure = dict(data=data, layout=layout_pie)
+    return figure
+
+
+def make_count_figure(well_statuses, well_types, year_slider):
+
+    layout_count = copy.deepcopy(layout)
+
+    dff = filter_dataframe(df, well_statuses, well_types, [1960, 2017])
+    g = dff[["API_WellNo", "Date_Well_Completed"]]
+    g.index = g["Date_Well_Completed"]
+    g = g.resample("A").count()
+
+    colors = []
+    for i in range(1960, 2018):
+        if i >= int(year_slider[0]) and i < int(year_slider[1]):
+            colors.append("rgb(123, 199, 255)")
+        else:
+            colors.append("rgba(123, 199, 255, 0.2)")
+
+    data = [
+        dict(
+            type="scatter",
+            mode="markers",
+            x=g.index,
+            y=g["API_WellNo"] / 2,
+            name="All Wells",
+            opacity=0,
+            hoverinfo="skip",
+        ),
+        dict(
+            type="bar",
+            x=g.index,
+            y=g["API_WellNo"],
+            name="All Wells",
+            marker=dict(color=colors),
+        ),
+    ]
+
+    layout_count["title"] = "Completed Wells/Year"
+    layout_count["dragmode"] = "select"
+    layout_count["showlegend"] = False
+    layout_count["autosize"] = True
+
+    figure = dict(data=data, layout=layout_count)
+    return figure
