+++ apps/dash-oil-and-gas/utils/helper_functions.py
@@ -1,0 +1,83 @@
+import datetime as dt
+from utils.data import points
+import math
+
+
+def human_format(num):
+    if num == 0:
+        return "0"
+
+    magnitude = int(math.log(num, 1000))
+    mantissa = str(int(num / (1000 ** magnitude)))
+    return mantissa + ["", "K", "M", "G", "T", "P"][magnitude]
+
+
+def filter_dataframe(df, well_statuses, well_types, year_slider):
+    dff = df[
+        df["Well_Status"].isin(well_statuses)
+        & df["Well_Type"].isin(well_types)
+        & (df["Date_Well_Completed"] > dt.datetime(year_slider[0], 1, 1))
+        & (df["Date_Well_Completed"] < dt.datetime(year_slider[1], 1, 1))
+    ]
+    return dff
+
+
+def produce_individual(api_well_num):
+    try:
+        points[api_well_num]
+    except:
+        return None, None, None, None
+
+    index = list(
+        range(min(points[api_well_num].keys()), max(points[api_well_num].keys()) + 1)
+    )
+    gas = []
+    oil = []
+    water = []
+
+    for year in index:
+        try:
+            gas.append(points[api_well_num][year]["Gas Produced, MCF"])
+        except:
+            gas.append(0)
+        try:
+            oil.append(points[api_well_num][year]["Oil Produced, bbl"])
+        except:
+            oil.append(0)
+        try:
+            water.append(points[api_well_num][year]["Water Produced, bbl"])
+        except:
+            water.append(0)
+
+    return index, gas, oil, water
+
+
+def produce_aggregate(selected, year_slider):
+
+    index = list(range(max(year_slider[0], 1985), 2016))
+    gas = []
+    oil = []
+    water = []
+
+    for year in index:
+        count_gas = 0
+        count_oil = 0
+        count_water = 0
+        for api_well_num in selected:
+            try:
+                count_gas += points[api_well_num][year]["Gas Produced, MCF"]
+            except:
+                pass
+            try:
+                count_oil += points[api_well_num][year]["Oil Produced, bbl"]
+            except:
+                pass
+            try:
+                count_water += points[api_well_num][year]["Water Produced, bbl"]
+            except:
+                pass
+        gas.append(count_gas)
+        oil.append(count_oil)
+        water.append(count_water)
+
+    return index, gas, oil, water
